# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "DamascusTransit API",
    "description": "Real-time transit tracking and fleet management \u2014 multi-tenant SaaS edition.\n\n## Authentication\n\nMost read endpoints are public but require an `?operator=<slug>` query parameter.\n1. **POST /api/v1/auth/login** \u2014 exchange email/password for a JWT token\n2. Include the token as `Authorization: Bearer <token>` on protected endpoints\n\nRoles: `super_admin` \u00b7 `admin` \u00b7 `dispatcher` \u00b7 `driver` \u00b7 `viewer`\n\nTokens expire after 24 hours.",
    "contact": {
      "name": "Damascus Transit System",
      "email": "support@damascustransit.com"
    },
    "license": {
      "name": "MIT"
    },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/traccar/position": {
      "post": {
        "tags": [
          "traccar"
        ],
        "summary": "Traccar Position Webhook",
        "description": "Webhook for Traccar GPS position updates. Secured by HMAC signature.",
        "operationId": "traccar_position_webhook_api_traccar_position_post",
        "parameters": [
          {
            "name": "x-traccar-signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "HMAC-SHA256 signature",
              "title": "X-Traccar-Signature"
            },
            "description": "HMAC-SHA256 signature"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TraccarPosition"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/traccar/event": {
      "post": {
        "tags": [
          "traccar"
        ],
        "summary": "Traccar Event Webhook",
        "description": "Webhook for Traccar events (engine on/off, speeding, etc). Secured by HMAC.",
        "operationId": "traccar_event_webhook_api_traccar_event_post",
        "parameters": [
          {
            "name": "x-traccar-signature",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "description": "HMAC-SHA256 signature",
              "title": "X-Traccar-Signature"
            },
            "description": "HMAC-SHA256 signature"
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TraccarEvent"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WebhookResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/push/subscribe": {
      "post": {
        "tags": [
          "push"
        ],
        "summary": "Subscribe Push",
        "description": "Store a Web Push subscription.\n\nPassengers associate subscriptions with stop IDs to receive arrival alerts.\nDrivers receive dispatch and route-change notifications.",
        "operationId": "subscribe_push_api_push_subscribe_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushSubscribeRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushSubscribeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "push"
        ],
        "summary": "Unsubscribe Push",
        "description": "Remove a Web Push subscription by endpoint URL.",
        "operationId": "unsubscribe_push_api_push_subscribe_delete",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": true,
                "type": "object",
                "title": "Req"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushUnsubscribeResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/push/broadcast": {
      "post": {
        "tags": [
          "push"
        ],
        "summary": "Broadcast Push",
        "description": "Broadcast a push notification to all matching subscribers (admin only).\n\nFilter by role ('passenger', 'driver', or None for all).\nReturns counts of sent/failed/skipped notifications.",
        "operationId": "broadcast_push_api_push_broadcast_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PushBroadcastRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushBroadcastResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/api/push/test": {
      "post": {
        "tags": [
          "push"
        ],
        "summary": "Test Push Self",
        "description": "Send a test push notification to all current admin subscriptions.",
        "operationId": "test_push_self_api_push_test_post",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PushTestResponse"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PushBroadcastRequest": {
        "properties": {
          "title": {
            "type": "string",
            "maxLength": 200,
            "minLength": 1,
            "title": "Title"
          },
          "body": {
            "type": "string",
            "maxLength": 1000,
            "minLength": 1,
            "title": "Body"
          },
          "icon": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Icon"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          },
          "data": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "title",
          "body"
        ],
        "title": "PushBroadcastRequest"
      },
      "PushBroadcastResponse": {
        "properties": {
          "sent": {
            "type": "integer",
            "title": "Sent"
          },
          "failed": {
            "type": "integer",
            "title": "Failed"
          },
          "skipped": {
            "type": "integer",
            "title": "Skipped"
          }
        },
        "type": "object",
        "required": [
          "sent",
          "failed",
          "skipped"
        ],
        "title": "PushBroadcastResponse"
      },
      "PushSubscribeRequest": {
        "properties": {
          "subscription": {
            "$ref": "#/components/schemas/PushSubscription"
          },
          "stopIds": {
            "anyOf": [
              {
                "items": {},
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Stopids"
          },
          "role": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "subscription"
        ],
        "title": "PushSubscribeRequest"
      },
      "PushSubscribeResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "endpoint": {
            "type": "string",
            "title": "Endpoint"
          },
          "role": {
            "type": "string",
            "title": "Role"
          }
        },
        "type": "object",
        "required": [
          "status",
          "endpoint",
          "role"
        ],
        "title": "PushSubscribeResponse"
      },
      "PushSubscription": {
        "properties": {
          "endpoint": {
            "type": "string",
            "title": "Endpoint"
          },
          "keys": {
            "additionalProperties": {
              "type": "string"
            },
            "type": "object",
            "title": "Keys"
          }
        },
        "type": "object",
        "required": [
          "endpoint",
          "keys"
        ],
        "title": "PushSubscription"
      },
      "PushTestResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Count"
          },
          "message": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Message"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "PushTestResponse"
      },
      "PushUnsubscribeResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "PushUnsubscribeResponse"
      },
...
      "TraccarEvent": {
        "properties": {
          "eventId": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Eventid"
          },
          "type": {
            "type": "string",
            "title": "Type"
          },
          "serverTime": {
            "type": "integer",
            "title": "Servertime"
          },
          "deviceId": {
            "type": "integer",
            "title": "Deviceid"
          },
          "deviceName": {
            "type": "string",
            "title": "Devicename"
          },
          "data": {
            "additionalProperties": true,
            "type": "object",
            "title": "Data"
          }
        },
        "type": "object",
        "required": [
          "type",
          "serverTime",
          "deviceId",
          "deviceName",
          "data"
        ],
        "title": "TraccarEvent"
      },
      "TraccarPosition": {
        "properties": {
          "deviceId": {
            "type": "integer",
            "title": "Deviceid"
          },
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          },
          "altitude": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Altitude"
          },
          "speed": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Speed"
          },
          "heading": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Heading"
          },
          "accuracy": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Accuracy"
          },
          "timestamp": {
            "type": "integer",
            "title": "Timestamp"
          }
        },
        "type": "object",
        "required": [
          "deviceId",
          "latitude",
          "longitude",
          "timestamp"
        ],
        "title": "TraccarPosition"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
      "WebhookResponse": {
        "properties": {
          "status": {
            "type": "string",
            "title": "Status"
          },
          "timestamp": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Timestamp"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "detail": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Detail"
          }
        },
        "type": "object",
        "required": [
          "status"
        ],
        "title": "WebhookResponse"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
...
}